Move utility classes to utils package It will get rid of update/metrics classes in main package
