+++ src/me/zombie_striker/qg/QAMain.java
@@ -12,16 +12,23 @@ import me.zombie_striker.qg.attachments.AttachmentBase;
 import me.zombie_striker.qg.config.*;
 import me.zombie_striker.qg.guns.*;
 import me.zombie_striker.qg.guns.projectiles.*;
+import me.zombie_striker.qg.guns.projectiles.HomingRocketProjectile;
+import me.zombie_striker.qg.guns.projectiles.MiniNukeProjectile;
+import me.zombie_striker.qg.guns.projectiles.RocketProjectile;
 import me.zombie_striker.qg.guns.utils.*;
 import me.zombie_striker.qg.handlers.*;
 import me.zombie_striker.qg.handlers.chargers.*;
 import me.zombie_striker.qg.handlers.reloaders.*;
 import me.zombie_striker.qg.miscitems.*;
+import me.zombie_striker.qg.miscitems.IronSightsToggleItem;
 import me.zombie_striker.qg.miscitems.ThrowableItems.ThrowableHolder;
 import me.zombie_striker.qg.npcs.Gunner;
 import me.zombie_striker.qg.npcs.GunnerTrait;
 import me.zombie_striker.qg.npcs_sentinel.SentinelQAHandler;
 
+import me.zombie_striker.qg.utils.GithubDependDownloader;
+import me.zombie_striker.qg.utils.GithubUpdater;
+import me.zombie_striker.qg.utils.Metrics;
 import org.bukkit.*;
 import org.bukkit.command.Command;
 import org.bukkit.command.CommandSender;
